fix: wire CleanupManager into disk commit for SIGKILL orphan cleanup
Move CleanupManager from core::signal to engine so both core and
transfer can share the same process-wide registry without introducing a
circular dependency. core::signal::CleanupManager is now a re-export.

The disk commit thread (process_file and process_whole_file) now
registers each temp file with CleanupManager after TempFileGuard
creation and unregisters it after a successful commit. This ensures
that a SIGKILL - which bypasses Drop - leaves a registry that a
subsequent startup or signal handler can sweep.

Only temp+rename paths are registered; inplace and device writes
operate on the final destination and need no orphan tracking.
